@@ -2,7 +2,8 @@ from locust import User
 from locust.exception import LocustError
 
 import time
-from typing import Any, Callable
+from collections.abc import Callable
+from typing import Any
 
 import grpc
 import grpc.experimental.gevent as grpc_gevent
