@@ -4,7 +4,7 @@ import collections
 import logging
 import os
 import sys
-from typing import Callable
+from collections.abc import Callable
 
 import gevent
 
